+++ src/Stormancer.Plugins/Blobs/Stormancer.Server.Plugins.BlobStorage/IBlobStorage.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -21,7 +22,7 @@ namespace Stormancer.Server.Plugins.BlobStorage
         /// <param name="content"></param>
         /// <param name="contentType"></param>
         /// <returns></returns>
-        ValueTask<CreateBlobResult> CreateBlobAsync(string blobStoreId,string path, ReadOnlyMemory<byte> content, string contentType);
+        ValueTask<CreateBlobResult> CreateBlobAsync(string blobStoreId, string path, ReadOnlyMemory<byte> content, string contentType);
 
         /// <summary>
         /// Deletes a blob
@@ -29,6 +30,13 @@ namespace Stormancer.Server.Plugins.BlobStorage
         /// <param name="path"></param>
         /// <returns></returns>
         ValueTask<DeleteBlobResult> DeleteAsync(string path);
+
+        /// <summary>
+        /// Gets the content of a blob.
+        /// </summary>
+        /// <param name="path"></param>
+        /// <returns></returns>
+        ValueTask<GetBlobContentResult> GetBlobContent(string path);
     }
 
     /// <summary>
@@ -39,7 +47,7 @@ namespace Stormancer.Server.Plugins.BlobStorage
         /// <summary>
         /// Gets or sets a value indicating if the operation was successful.
         /// </summary>
-        [MemberNotNullWhen(true,"Path")]
+        [MemberNotNullWhen(true, "Path")]
         public bool Success { get; set; }
 
         /// <summary>
@@ -59,12 +67,47 @@ namespace Stormancer.Server.Plugins.BlobStorage
         public bool Success { get; set; }
     }
 
+    /// <summary>
+    /// Result of a get blob request.
+    /// </summary>
+    public class GetBlobContentResult : IDisposable
+    {
+        /// <summary>
+        /// Gets or sets a value indicating if the operation was successful.
+        /// </summary>
+        [MemberNotNullWhen(true, "Content")]
+        [MemberNotNullWhen(true,"ContentType")]
+        [MemberNotNullWhen(false,"Reason")]
+        public required bool Success { get; init; }
+
+        /// <summary>
+        /// Gets a <see cref="System.IO.Stream"/> object exposing the blob content.
+        /// </summary>
+        public Stream? Content { get; init; }
+
+        /// <summary>
+        /// Gets a <see cref="string"/> providing a reason for the failure.
+        /// </summary>
+        public string? Reason { get; init; }
+
+        /// <summary>
+        /// Gets the content type of the blob.
+        /// </summary>
+        public string? ContentType { get; init; }
+
+        /// <inheritdoc/>
+        public void Dispose()
+        {
+            Content?.Dispose();
+        }
+    }
+
     internal class BlobStorage : IBlobStorage
     {
         private readonly IEnumerable<IBlobStorageBackend> _backends;
         private readonly IConfiguration _configuration;
 
-        public BlobStorage(IEnumerable<IBlobStorageBackend> backends, IConfiguration configuration )
+        public BlobStorage(IEnumerable<IBlobStorageBackend> backends, IConfiguration configuration)
         {
             _backends = backends;
             _configuration = configuration;
@@ -73,12 +116,12 @@ namespace Stormancer.Server.Plugins.BlobStorage
         {
             var config = _configuration.GetValue<BlobStorageConfigurationSection>(BlobStorageConfigurationSection.PATH) ?? new BlobStorageConfigurationSection();
 
-            if(!config.BlobStores.TryGetValue(blobStoreId, out var storeConfiguration))
+            if (!config.BlobStores.TryGetValue(blobStoreId, out var storeConfiguration))
             {
                 return new CreateBlobResult { Success = false };
             }
 
-            foreach(var backend in _backends)
+            foreach (var backend in _backends)
             {
                 if (backend.CanHandle(storeConfiguration))
                 {
@@ -101,15 +144,15 @@ namespace Stormancer.Server.Plugins.BlobStorage
         {
             var firstSlash = path.IndexOf('/');
 
-            if(firstSlash > -1)
+            if (firstSlash > -1)
             {
-                return (path.Substring(0,firstSlash), path.Substring(firstSlash + 1));
+                return (path.Substring(0, firstSlash), path.Substring(firstSlash + 1));
             }
             else
             {
-                throw new ArgumentException($"'{path}' is not a valid blob path",nameof(path));
+                throw new ArgumentException($"'{path}' is not a valid blob path", nameof(path));
             }
-            
+
         }
 
         private string CreatePath(string blobStoreId, string path)
@@ -128,16 +171,38 @@ namespace Stormancer.Server.Plugins.BlobStorage
                 return ValueTask.FromResult(new DeleteBlobResult { Success = false });
             }
 
-            foreach(var backend in _backends)
+            foreach (var backend in _backends)
             {
-                if(backend.CanHandle(storeConfiguration))
+                if (backend.CanHandle(storeConfiguration))
                 {
-                    return backend.DeleteAsync(storeConfiguration,blobPath);
+                    return backend.DeleteAsync(storeConfiguration, blobPath);
                 }
             }
 
             return ValueTask.FromResult(new DeleteBlobResult { Success = false });
 
+        }
+
+        public ValueTask<GetBlobContentResult> GetBlobContent(string path)
+        {
+            var config = _configuration.GetValue<BlobStorageConfigurationSection>(BlobStorageConfigurationSection.PATH) ?? new BlobStorageConfigurationSection();
+
+            var (blobStoreId, blobPath) = ParsePath(path);
+
+            if (!config.BlobStores.TryGetValue(blobStoreId, out var storeConfiguration))
+            {
+                return ValueTask.FromResult(new GetBlobContentResult { Success = false, Reason = $"blobStoreNotFound?id={blobStoreId}" });
+            }
+
+            foreach (var backend in _backends)
+            {
+                if (backend.CanHandle(storeConfiguration))
+                {
+                    return backend.GetContentAsync(storeConfiguration, blobPath);
+                }
+            }
+
+            return ValueTask.FromResult(new GetBlobContentResult { Success = false, Reason = $"blobStoreBackendNotFound?id={blobStoreId}" });
         }
     }